backend/requirements.txt
Flask
requests
numpy
faiss-cpu
sentence-transformers
python-dotenv